test sucessfull without intrinsic scatter and abble to install spp
