+++ sugar_plus_plus/__init__.py
@@ -6,3 +6,5 @@ Some description.
 
 from .toolbox import distance_modulus
 from .toolbox import weighted_avg_and_std
+
+from .hd import hubble_diagram
